Fix MIDI receive on DICE (Re: #330)
Code more or less copied from MOTU. It seems MIDI bytes in the input
buffer might occur at all locations, not just 8-sample-aligned, so we'll
have to step through each incoming quadlet and see if it's MIDI.

On the jackd side, we'll have to obey the 8-sample-alignment rule.

An upcoming patch will indent the code, just for reviewing, I decided to
have no whitespace fixes.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@1973 2be59082-3212-0410-8809-b0798e1608f0