DICE: Use nicknames instead of GUID
Change all channel names to use a device's nickname. Historically, we've
used the GUID to distinguish different devices on the bus.

In case of colliding nicknames, we'd fail to register the jack port, so
owners of multiple devices have to make sure to use different nicknames.

Patch by Philippe Carriere

git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2059 2be59082-3212-0410-8809-b0798e1608f0