DICE-EAP Corrects the number of router entries for clearing in EAP::RouterConfig::write
By Philippe Carriere:

EAP::RouterConfig::write assumes the router has 128 entries, something
which is not part of the Dice EAP specifications.  This number is
replaced by the maximal number of entries as read in the rom; introduce
the function getMaxNbRouterEntries() as a public function of EAP. Then
use it in EAP::RouterConfig::write.

git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2092 2be59082-3212-0410-8809-b0798e1608f0