audiofire: correct reporting of sample rate for device firmware 5.8.  Patch (and the notes below) are from Takashi Sakamoto.
This is a patch for an issue which is not reported. I confirmed this issue with AudioFirePre8 and AudioFire4.

With firmware version 5.8, FFADO mixer fails to set sampling rate and clock source. This is related to the behaviour against EfcGetClock command just after EfcSetClock command.

With previous firmware version, the device can return the actual value with EfcGetClock command just after EfcSetClock. But with this version, there needs to be approximately 100msec. During this period, EfcGetClock return the PREVIOUS state of both of sampling rate and clock source.

This patch adds sleep during 150msec after deserializing response of EfcSetClock.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2391 2be59082-3212-0410-8809-b0798e1608f0