[libavc] complement 'Processing function block' command
Current implement of libavc is lack of actual work for 'Processing function block' command defined in 'AV/C Audio Subunit Specification 1.0 (1394TA Document 1999008)'. I guess the reason is there are several types of the command. To implement all of the commands in the specification is too cost.

But here, a driver for BeBoB/MAudio devices needs this command to control device's internal mixer. These devices can be controlled by either 'single' or 'multiple' type of the command.

For these reasons, this commit add 'single type' of 'processing function block command'.

git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2442 2be59082-3212-0410-8809-b0798e1608f0