[BeBoB/MAudio] Add PreSonus::FireboxDevice class to support functionality to switch clock source
There is a difference of CCM command between GenericAVC and FIREBOX. In FIREBOX, the most significant bit of Plug ID always stands. This is also confirmed in M-Audio Firewire 410/Firewire Audiophile/Firewire Solo. I guess there is BeBoB specific extension of CCM but I have no way to confirm it.

In this reason, this commit adds PreSonus::FireboxDevice.

git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2466 2be59082-3212-0410-8809-b0798e1608f0