fireworks: Partial revert of r2452, remote several trials of EfcOverAVC
@2452, this trials were added for firmware version 5.0 or later of AudioFire12.
When adding it, the failure related some commands are considered due to command
failure so retry is reasonable.

But according to additional investigation, it's due to wrong contents of these
commands. So command are itself successful.

To give better experiences to users, these retries should removed and go to
fallback as soon as possible.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2509 2be59082-3212-0410-8809-b0798e1608f0