fireworks: Improve a workaround for AudioFire12 with recent firmwares
Firmware version 5.0 or later for AudioFire12 has quirks. One of the quirks is
to return invalid values in content of response against command of polled
values.

Acording to ticket #360, this causes assertion which prevents from
buffer-over-run:
http://subversion.ffado.org/ticket/360

Revision 2451 removes this assertion and adds a prevention of buffer-over-run.
http://subversion.ffado.org/changeset/2451

But this is not enough as a workaround because it returns false.

This commit allows to continue process after compensating these values.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2514 2be59082-3212-0410-8809-b0798e1608f0