fireworks: Fix a bug of fallback mechanism for clock information
I'm working with wrong understanding about #360.

The log means some commands are successful but they includes wrong values.
In my previous work, I just tested a case that these commands are failed.
So this issue is not still resolved.

This commit adds Fireworks::Device::getClock() with fallback mechanism and
all of methods which get clock information call this method. This fallback
mechanism checks values in contents of response and correct the values if
they're invalid.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2519 2be59082-3212-0410-8809-b0798e1608f0