From Hector Martin <marcan@marcan.st>
I got an ADAT device to test this input and it turns out I had botched
the input selection. The ADAT port goes into the second ADAT receiver,
not the first. This makes sense, as S/PDIF via ADAT is on the fourth
AES receiver, which shares the same physical pin on the DICE chip.

Signed-off-by: Hector Martin <marcan@marcan.st>



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2606 2be59082-3212-0410-8809-b0798e1608f0