support/dbus/: work around a bug in dbus-c++.
dbus-c++ 0.9.0 and earlier has a bug which prevents compilation under gcc7
(and probably later versions).  If DBUS_HAS_RECURSIVE_MUTEX is not defined
the compiler identifies a type mismatch between the MutexFreeFn function
typedef and the implemented Threading::mutex_free() method through the
_init_threading() call.  DBUS_HAS_RECURSIVE_MUTEX is defined in dbus-c++'s
internal config.h but this is not installed for system wide use and is not
referenced in any of the dbus-c++ include files.

DBUS_HAS_RECURSIVE_MUTEX is defined in dbus-c++'s config.h for dbus 0.95 or
later.  The bool variation of the mutex functions are deprecated, and since
dbus-0.95 dates from 2007 it is safe to assume that everyone is using the
void versions.  Therefore the simplistic approach of defining
DBUS_HAS_RECURSIVE_MUTEX if it is not already defined prior to including
dbus-c++/dbus.h should be safe.  In addition it is necessary to do this
before including any of the *-glue.h files.  These include dbus-c++/dbus.h
but because they are generated it's not easy to insert the workaround code
into them.

Thanks to Orcan Ogetbil for the initial report, looking into and identifying
the true source of the problem in dbus-c++, and proposing the solution.  The
full discussion was on the FFADO-devel list, subject "gcc7 build fixes",
14-16 February 2017.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2674 2be59082-3212-0410-8809-b0798e1608f0