python: replace all "print ..." statements with "print( ... )", which is compatible with current versions of python2 and python3.  The replacement was done with a straight-forward shell command:
  for i in $(find . -name \*.py) ; do 
    sed -i.bak 's/print \(.*\)/print( \1 )/' $i ; 
  done

This will fail if any of the print statements extend beyond a single line. 
Inspection suggests there are none of these, but there is always a chance
that one or two have been missed.  If so they can obviously be fixed up
later.

The move to the new "print" syntax was included in Xavier Forestier's
November 2016 patch set.  The majority of the changes needed to make FFADO
run under python3 involved fixing the "print" syntax.  Other issues
identified in that patch set will be dealt with in future patches.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2680 2be59082-3212-0410-8809-b0798e1608f0