ffado-mixer: Centralise imports of PyQt modules and make compatible with PyQt5.
In Xavier Forestier's patch set submitted to ffado-devel in November 2016,
changes were required to every PyQt import statement if PyQt5 were to to be
used instead.  Since the current intent is to retain compatiblity to both
PyQt4 and PyQt5 a slightly different approach was needed.

All PyQt module imports are now handled by the newly added import_pyqt
module.  This transparently loads PyQt4 or PyQt5 modules depending on what
is found on the system at runtime.  Presently PyQt4 is tried first with a
fallback to PyQt5 if PyQt4 is not found, but this could change in future. 
The result is that the ffado-mixer code should now be very close to being
compatible with PyQt5 without additional patches.  Having said that, some
build system changes are still needed to make PyQt5 an option, along with
some other minor fixes.

This is perhaps not the neatest way of achieving the desired result, but it
seems to work.  It avoids the need to patch every ffado-mixer source file
depending on the PyQt version in use, and allows ffado-mixer to adapt to the
PyQt version found at runtime.

The import_pyqt module also sets two global version variables:
ffado_pyqt_version (the version of PyQt in use) and ffado_python3 (true if
python3 is in use).  These may be of use when addressing remaining
PtQt4/PyQt5 and Python 2/3 issues.

This patch has been tested with PyQt4 under Python 2.7.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2694 2be59082-3212-0410-8809-b0798e1608f0