ffado_mixer: add a "Themes" submenu to the "View" menu.
Xavier Forestier's November 2016 patchset included code to add themes under
the "Help" menu which could be selected by the user.  The code to do this is
relatively simple so there seems little reason to hold it back.  The only
change compared to the submitted patch is to move the themes to a submenu of
the "View" menu since conceptually this is a better place than the top of
the "Help" menu.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2700 2be59082-3212-0410-8809-b0798e1608f0