ffado-mixer: add device mixers to "View" menu.
Xavier Forestier's November 2016 patchset included code which added an entry
for each active device's mixer to the "View" menu.  For users with a number
of different devices this may prove useful.  This commits Xavier's original
patch was with some minor changes to ensure the guid key is resolvable by
setTabVisible() under PyQt4.  It has not been tested under PyQt5.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2701 2be59082-3212-0410-8809-b0798e1608f0