ffado_mixer: use graphical icon instead of name for mixer pane if available.
From Xavier Forestier's November 2016 patchset.  In amongst the Qt5 changes
was this code to utilise a device-specific graphical image instead of the
textual device model name in the global mixer pane.  The only change from
Xavier's patch is the omission of a statement to set the nickname
placeholder text to the name of the device.  The nickname field is present
to support a particular feature of some interfaces, so it shouldn't really
be repurposed like this.  Otherwise there is a high likelihood of it
introducing confusion when used with devices which support the nickname
feature.

FFADO does not currently ship with any device graphics.  Contributions are
welcome but care will be needed to ensure that copyrighted images from
vendors are not used.  The use of official images could give an incorrect
impression that ffado-mixer is officially sanctioned by the vendors which is
not generally the case.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2702 2be59082-3212-0410-8809-b0798e1608f0