mackieonyx: fix mixer module name inconsistency.
The name of the python module implementing the Mackie Onyx mixer
("mackie_onyxmixer") was inconsistent with the class name as defined in the
"configuration" file ("MackieOnyx").  The class as defined in
"mackie_onyxmixer" ("MackieOnyx") was fine though.  To avoid causing
unnecessary churn in existing "configuration" files, the mixer module was
renamed to "mackieonyx", thus matching what is expected in the
"configuration" file definition.

This name mismatch was spotted by Xavier Forestier, who implemented a
slightly different fix in his November 2016 patch set.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2703 2be59082-3212-0410-8809-b0798e1608f0