saffire_dice: allow for trailing space in ConfigROM model name.
Under certain conditions, the model name returned by configrom.getModelName()
can include trailing spaces.  There is a vague recollection that a change
in the kernel or libraw1394 at some point in the past might give rise to
this, but no reference can be immediately tracked down.  To address this,
simply use the rstrip() method to remove trailing white space.  This ensures
that the code continues to work with or without the trailing white space.

This issue caused ffado-mixer to exit with an error complaining that
GlobalMute was not defined for the Saffire mixer widget.  Due to the
unexpected white space, the device name never matched any of the expected
names and therefore the appropriate UI file (which created the GlobalMute
objects, among many others) was never loaded.

Thanks to Dominik Meister for reporting the problem, running further tests
to confirm the fundamental cause and suggesting the fix.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2708 2be59082-3212-0410-8809-b0798e1608f0