Additional scons3 compatibility fixes (add implicit bytes/string conversions).
Within the scons scripts, some of the env[VAR] calls return bytes under
scons3.  Implicit conversion of bytes to strings has been removed in python3
which necessitates the use of decode() in some instances which expect
strings to prevent type mismatch errors.  This change does not affect
compatibility with scons2.

Thanks to Orcan Ogetbil for his scons3 tests and the patch (via the
ffado-devel mailing list).


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2718 2be59082-3212-0410-8809-b0798e1608f0