ffado-mixer.appdata.xml: fix project licence identifier.
While the table at http://spdx.org/licenses/ indicates that "GLP-2.0-only"
is the appropriate tag for GPLv2 software, an alternative specification at
https://spdx.org/spdx-specification-21-web-version#h.1jlao46 says otherwise.
The "appstream-util validate" command on Fedora reports "GPL-2.0-only" as
invalid, so we'll run with "GPL-2.0" instead.

Thanks to Orcan Ogetbil for running the appstream-util on his system and
pointing out the URL with the alternative SPDX licence identifier
definitions.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2757 2be59082-3212-0410-8809-b0798e1608f0