ffado-diag: correct regex used to find firewire controllers in lspci output.
The previous re.findall() regex expected the PCI bus ID to be within literal
parentheses, while the intent was probably to specify a subexpression (or
"group" as python calls it).

When a subexpression (group) is included in the regex passed to
re.findall(), the values produced re.findall() are the contents of the
subexpression.  A call to the group() method is not required.

These changes have been tested on python2.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2763 2be59082-3212-0410-8809-b0798e1608f0