SConstruct: revise for recent python3 environments.
This mostly reverts the changes made in r2715 and r2716 on 6 Nov 2017.  At
that time the scons3/python3 combination were throwing TypeError errors if
the "b" prefix was not present in the split() call.  Recent reports from
David Runge and Orcan Ogetbil indicate that not only is the "b" prefix not
needed anymore, but it causes an error.

The content of this patch was supplied by David Runge who tested it under
python 3.5.7, 3.6.8 and 3.7.3.  Orcan Ogetbil later verified it in his
scons3/python3 environment.  It also remains functional under
scons2/python2.  On the basis of this it seems safe to commit the change.

Refer to the "Documentation build options (was: Re: Scons3 compatibility)"
thread on the ffado-devel mailing list thread for the full discussion.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2777 2be59082-3212-0410-8809-b0798e1608f0