ffado-mixer: disable the optional registration prompt.
While usage statistics were useful in the past when discussing Linux support
with vendors, in 2020 this is no longer the case.  Vendors are generally no
longer shipping Firewire-based products so these statistics are not of
practical value anymore.  Since they are no longer used there is little
point to collect them, so disable the registration dialog box.

While this change has been pondered for a while, it has been brought to a
head by the changes to the website infrastructure in 2020.  To continue to
support the device registration, the website services used by the
registration code would need to be supported.  Given the limited utility of
the data produced, the time required to do this does not seem worthwhile.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2791 2be59082-3212-0410-8809-b0798e1608f0