Update displayed copyright year range where relevant.
There is little doubt that the copyright year range of some components of
FFADO have not been kept up to date.  For those where this can be clearly
demonstrated, update the range to include the current year (2021).  Others
can be addressed in future once it is determined that they were affected by
later changes.  There are certainly some FFADO utilities not linked against
libffado which have not received code changes since the years currently
claimed.

Thanks to Pander (via the ffado-devel list) for drawing our attention to
this issue.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2796 2be59082-3212-0410-8809-b0798e1608f0