Use capital "F" for "Firewire" in non-code settings.
Since "Firewire" is a proper noun it should technically be started with a
capital letter.  Thanks to Pander on the ffado-devel mailing list for
providing the patch.

When discussing the "firewire" backend of JACK a capitial is not used
because the respective command line option of jackd is "firewire", not
"Firewire".


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2798 2be59082-3212-0410-8809-b0798e1608f0