Cosmetic: "Firewire" becomes "FireWire".
Officially both the "F" and "W" were capitalised in the FireWire name, so
reflect this throughout FFADO's source tree.  This mostly affects comments.

This patch originated from pander on the ffado-devel mailing list.  To
maintain consistency, the committed version has been expanded to include
files not originally included in the original patch.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2802 2be59082-3212-0410-8809-b0798e1608f0