ffado-mixer: default to dark theme, save and restore the user's theme choice.
Dark themes are often preferred when working with audio software.  If the
"ukui-dark" theme is present on the system, use this as ffado-mixer's
default theme.

While ffado-mixer has permitted the user to choose a theme for some time,
this choice was not recalled when ffado-mixer was next started.  Avoid this
obvious inconvenience for cases where the default theme is not the user's
preference.

This functionality was contributed by "pander" via the ffado-devel mailing
list.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2805 2be59082-3212-0410-8809-b0798e1608f0