Cosmetic: "firewire" should be "FireWire" when used as the bus name.
Similarly to r2802 and r2810, "FireWire" should be used when referring to
the name of the bus.  This patch corrects this throughout the source tree
for completeness.  While there are a small number of mostly debug output
strings affected, most of the changes are to comments or developer documents
where they are of little consequence.  Thanks to Pander who suggested the
need to look into this on the ffado-devel mailing list.

At least in theory, remaining instances of "firewire" in the source tree
should remain as they are because they refer to case-sensitive identifiers
defined externally (such as the "firewire" jackd backend name).


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2811 2be59082-3212-0410-8809-b0798e1608f0