deb/: add "qt5-ukui-platformtheme" as a recommendation.
Some have found the ukui dark theme to be a good match for ffado-mixer,
so add it as a recommendation in the dev SConscript.  This was suggested by
Pander on the ffado-devel mailing list.  This file isn't used to produce
official Debian packages, but acts as a starting point for developers who
would like to create their own packages.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2815 2be59082-3212-0410-8809-b0798e1608f0