SConstruct: attempt to address an inconsistency with conf.Finish().
In March 2020, a discussion on FFADO-devel (subject: "FFADO-mixer and
python") identified that setting "env" to the return value of the
conf.Finish() function was not correct.  It just so happens that it worked
up to and including scons 3.0.4.  In scons 3.0.5 the behaviour of
conf.Finish() changed, whereby any keys added to the "env" object were no
longer preserved in the environment returned by conf.Finish() if the
"--config=force" scons option was specified.

At the time of this discussion no consensus was reached on the correct
course of action.  The primary concern was to avoid fixing some systems
while breaking others.

This problem has now affected a second user as reported on the FFADO-user
mailing list (subject: Profire2626 routing issues).  This patch implements
one of the possible options floated in March 2020 - to simply discard the
return value of the conf.Finish() function.  If it turns out that this
breaks a significant number of systems an alternative solution will need to
be found.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2816 2be59082-3212-0410-8809-b0798e1608f0