ControlInterface::getContignuous(): explicitly return an integer type.
Code which calls this function treats the return type from getContignuous()
as an integer.  However, this only works if explicit type conversion is in
place because the underlying interface returns a float.  In python 3.10
and/or later versions of pyqt 5 this is not the case.  To avoid the need to
make changes to every caller of getContignuous() it seems best to alter
getContignuous() itself so an integer is explicitly returned.

Thanks to Filippo Bardelli for pointing out the problem on the ffado-devel
list, providing an initial patch and for testing the suggested fix.


git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2817 2be59082-3212-0410-8809-b0798e1608f0