SConstruct: fall back to argp-standalone if argp is not in the system libc.
While glibc provides a builtin implementation of argp, most other libcs do
not.  Fall back to using the argp-standalone library on systems where argp
is not part of the system libc.

This patch is based on an idea outlined on the ffado-devel mailing list by
Alyssa Ross.  As of the commit time, it has not been tested on a non-glibc
system.  It is anticipated that this patch will allow FFADO to be built
against musl libc without requiring the manual addition of argp-standalone
to the list of libraries to link against.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2848 2be59082-3212-0410-8809-b0798e1608f0