SConstruct: add a DATADIR option and use it in line with GNU conventions.
The new DATADIR option operates in the same way that datadir does as per the
GNU conventions.  By default it is used to define SHAREDIR, MANDIR and the
location of the metainfo/ directory for storing the ffado-mixer appdata
file.  This should make life easier for package maintainers who have to deal
with modified filesystem structures, allowing them to use the same methods
that they employ for other packages.

This change does shift the default MANDIR from ${PREFIX}/man/ to
${PREFIX}/share/man/.  The alternative location (which appears to be more
common) should not cause any issues, but if any unexpected side-effects are
observed please get in touch via the ffado-devel mailing list.

Thanks to Jan Tojnar for the suggestion and the patch.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2851 2be59082-3212-0410-8809-b0798e1608f0