[PATCH 3/3] Install XDG files manually.
Patch 3 of 3 from Jan Tojnar: I also got rid of the xdg-utils usage, which
simplifies the build script and will allow distros to install the XDG files
through scons.

xdg-utils are barely maintained, an overkill for our purposes, and do not
even support DESTDIR.

Installing the icon as hi64-apps-ffado instead of just ffado, since that is
actually listed in the desktop file.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2854 2be59082-3212-0410-8809-b0798e1608f0