crossbarrouter: handle "disconnected" state when restoring saved configuration.
Destinations in the crossbar router can be disconnected.  In other words, a
destination does not necessarily have a source selected.  When this is the
case, the saved state will list the destination without any source.  This
condition must be explicitly accounted for when the saved state is restored
because there won't be a second element in the output from split(). 
Furthermore, the disconnected state must be explicitly set so the intent
of the saved state takes effect no matter how the device might have been set
up leading into the state restoration.

Thanks to Nick Sorenson for noticing that the saved crossbar router state
wasn't being restored, and for testing the fix.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2871 2be59082-3212-0410-8809-b0798e1608f0