configrom: try to make MAudio ProFire 2626 GUIDs unique.
For unknown reasons, every ProFire 2626 has the same GUID.  Since FFADO uses
the GUID to identify individual devices, this means that FFADO will only
recognise and work with one ProFire 2626 even when more than one is
connected to the bus.

A reasonably complete exploration of this issue is detailed in the "Dual
ProFire 2626, both with the same GUID, only one interface used" thread on
the ffado-user mailing list in August 2018 [1].

To crudely work around this issue, add the node ID to the GUID when the
ProFire 2626 GUID is seen.  On typical systems this should make it possible
to use more than one ProFire 2626 at once through FFADO.  This probably
won't work if different ProFire 2626s are connected to distinct FireWire
buses since each bus enumerates its own node IDs starting at 0.  However, in
practice this would be a very rare situation, and therefore at this point in
time there is little benefit in trying to address it.

This patch has not been runtime tested due to the lack of access to a
ProFire 2626.  If it proves to cause problems it should be reverted.

[1] https://sourceforge.net/p/ffado/mailman/ffado-user/thread/2a5c131d-82e2-f94a-8e0b-e597005fdc15%40settrans.net/



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2872 2be59082-3212-0410-8809-b0798e1608f0