ffado-mixer: Address b'...' in "About FFADO" dialog under newer python versions
The FFADO version reported in the ffado-mixer "About FFADO" dialog is
obtained by calling into the libffado C function ffaod_get_version().  This
returns a "char *" which maps to the python byte string type.  Up until a
certain version of python, the string was rendered as expected into the
QMessageBox.  Python 3.9 is definitely okay, but other later versions may be
too.

In a python version later than 3.9, the FFADO version line was rendered
as

    b'libffado ...'

instead.  It is suspected that this is due to changes in the way strings and
byte strings are handled in later versions of python.

This patch attempts to address this by explicitly converting the returned
byte string into a string object.

Thanks to Pander for drawing our attention to this issue.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2874 2be59082-3212-0410-8809-b0798e1608f0