matrixmixer: fix a float/int type mismatch when indexing an array.
Recent versions of python are pickier when it comes to the type of array
indices.  This particular occurance in the code associated with reading and
applying a setup file was evidently missed in the earlier fixes for this
type of problem.  Thanks to Dmitry Sushkov for spotting the problem and
providing the fix.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2876 2be59082-3212-0410-8809-b0798e1608f0