Profire 2626: fix labelling of "Mixer/Out:01" in matrix mixer.
The matrix mixer displayed a label of "ADAT B/In:08" for the first column,
rather than "Mixer/Out:01".  It appears this is due to a typo in the
initial index number used when setting up the ADAT B/In channels.  This
caused "ADAT B/In:08" and "Mixer/Out:01" to have the same internal ID.

Thanks to Dmitry Sushkov for spotting the problem and providing the fix.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2878 2be59082-3212-0410-8809-b0798e1608f0