Profire 2626: avoid spaces in router source and destination names.
The presence of spaces in the names of channels used by the router caused
trouble when saving and restoring the router settings because they are used
as array keys.  As a result, while mixer settings could be restored, router
settings could not be.  Since this issue affects only the ADAT channels of
the Profire 2626, the easiest solution is to replace the space character in
the affected channel names.

This change affects the names of ADAT channels in the profire 2626 mixer, so
ADAT-related mixer channels may not restore correctly from a previously
created configuration file.  A fresh configuration file will need to be
created, or an existing one edited to replace the space in the ADAT channel
names with a hyphen ("-").

Thanks to Dmitry Sushkov for spotting this issue and providing a patch.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2879 2be59082-3212-0410-8809-b0798e1608f0