panelmanager: fix out-by-one index when restoring settings.
When a device's settings have been restored, its respective lines in the
configuration content are deleted so the next device can be processed. 
However, an out-by-one error meant that a device's closing "</device>" tag
was not deleted.  If more than one device's settings were present in the
configuration being restored, the next iteration of the device loop would
throw an error due to the errant "</device>" tag.

This comes about because python's

    del a[m:n]

construct deletes zero-based indices m to n-1 from the array "a", leaving
index n present.  In the case of the PanelManger::readSettings() method,
idxb is the index of the opening "<device>" tag while idxe is that of the
closing "</device>" tag.  Therefore all indices inclusively between idxb and
idxe must be removed.  If using "del", this therefore requires a range
specification of "idxb:idxe+1".

Thanks to Dmitry Sushkov for spotting this issue and suggesting the fix.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2882 2be59082-3212-0410-8809-b0798e1608f0