ffado-mixer: introduce preferences dialog with log enable option.
Up to this point ffado-mixer has included only a small number of settings
relating to the user interface which could be conveniently handled by items
in the "View" menu.  However, there are plans to introduce additional
settings which will be more conveniently accommodated in a preferences
dialog window.

This patch adds a preferences dialog along with a single initial control: a
checkbox to enable or disable the log display.

Thanks to Dmitry Sushkov for the patch.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2884 2be59082-3212-0410-8809-b0798e1608f0