ffado-mixer: improve error handling if restoring settings when streaming active
If streaming is active it is not possible to alter the clock source or
sampling rate.  Since these parameters are included in the saved state,
attempts to restore a saved state with streaming active would result in a
warning.  However, often the state being restored will contain the same
clock source and sample rate that the device is already configured with.  If
this is the case there is no reason to warn the user that these could not be
set even if streaming is active.

Thanks to Dmitry Sushkov for the patch.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2885 2be59082-3212-0410-8809-b0798e1608f0