ffado-mixer: further fix for Nickname handling.
r2877 attempted to fix the handling of the Nickname control which previously
referenced non-existing methods.  The choice of ascii() was, on balance, not
the best.  Firstly, it isn't available in python2.  However, more
importantly, it adds literal single quotes to the string it returns.  A
better option is str(), which is used elsewhere in the code base and should
to the right thing for the Nickname too.

Thanks to Dmitry Sushkov for reporting the problem.



git-svn-id: http://subversion.ffado.org/ffado/trunk/libffado@2886 2be59082-3212-0410-8809-b0798e1608f0